Initial setup of models
There is still some stuff to work on; for example, on the admin, when you select
an artist it should switch the album field so that it is only displaying albums
by the selected artist.

Also, album art has not been implemented yet for the Album model, mainly because
I haven't set up static content hosting yet.