Fixed up song template
After some soul-searching, decided to go with <table>. Arguably a lyricsheet is
a form of tabular data, but really it's for reasons of expediency in terms of
getting the stanzas to line up across columns.

Anyway, even if I do choose to switch to div + p later, it's not too difficult
because the context data was written so that it's striped in row-major order
when sent to the template. So it should be doable with some sort of grid-based
CSS framework.